Correcting placement of quickstarts tests
